Callable filter in column options
Proc is called with a single argument klass allowing for
producing a refining filter specific to data prefiltered in a controller.

One may also pass TabulatrData class member as a filter
either using #member or as a symbol,
if similar filtration is to be applied to several columns.
In that case, column is passed as the second argument to such function.